lk_tps65185: fix irq handling during suspend
Only free the irqs after the delayed-work check has run. If the irqs are freed
earlier and the suspend is aborted due to pending work this would lead to a
double free at the next suspend, as the matching resume function is only called
if the suspend function was sucessful.