disable reporting of key states during resume
The gpio-keys driver reports the state of all its keys during resume
and normally the input layer should filter out duplicate events of the
same state on its own (press when already pressed) but this does not
seem to work in our case.

As the press of the homebutton nevertheless always triggers the interrupt
and thus reaches the report-stage as well, we can disable the resume-time
reporting for our use case, to catch the double-report of the key.