maybe fix homekey issue
it seems the home-key-interrupt can happen after the interrupts already
got disabled when suspending - for whatever reason. This seems to throw
the system in for a loop, as the interrupt is marked as pending and thus
lets the system reattempt it all the time - especially during further suspends.

As the lock of the irq is then probably contented this will never get
resolved, making the system unable to go to suspend anymore - returning a
"busy" error code.

Try to resolve this in a hacky way by not marking the home-key interrupt as
pending in this case. It won't get reattempted and it was able to go to sleep
for me after the original issue happened.