lk_tps65185: fix race condition between _pwrdwn_work_func and _chg_mode
When _chg_mode is called it takes the lock and tries to cancel pending
_pwrdwn_work_func calls. The is a constellation, with _pwrdwn_work_func
already running and waiting for the lock, that is already held by _chg_mode.
Here cancel_delayed_work_sync will wait for _pwrdwn_work_func to finish
which will never happen, as _chg_mode will never release the lock.

Secondly add pm_wakeup calls in _pwrdwn_work_func, to make the sure the
system does not suspend while pwrdwn_work_func is running, as this could
also lead to a deadlock. Tasks get frozen first when suspending, so when
_pwrdwn_work_func is running at the time, it is possible it will be stopped
and keep holding the lock, making the later _chg_mode calls during suspend
hang the system, waiting for the lock.